Feat: add swagger documentation

+++ apps/nutritionist/src/app/auth/service/auth.service.ts
@@ -113,6 +113,14 @@ export class AuthService {
    * @returns A promise that resolves when the sign-out process is complete.
    */
   async signOut(id: string): Promise<void> {
+    const account = await this.repository.findAccountById(id);
+
+    if (!account.fcmToken && !account.refreshToken) {
+      throw new UnauthorizedException(
+        AuthErrorMessage.ERR_ACCOUNT_ALREADY_SIGN_OUT,
+      );
+    }
+
     await this.appJwtService.deleteRefreshToken(id);
 
     await this.repository
